Syntax changes
If you are starting a new label which is a new sticky note in Twine. It needs to end in : so an example is label Bear: .
The other thing is menu can't have flavor text. It can only be the options to choose and the text it has in it.