Added System for Accessing Branches
Created a checking system in the first scene to confirm which branches have been completed. So when the player completes a branch, it will not be accessible again, this way they progress to a new branch.

Also cleaned up the name variables for the main cast of characters and minor characters.